resurrect 64bit compile
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2229 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ RenderDevice.cpp
@@ -1136,12 +1136,12 @@ IndexBuffer* RenderDevice::CreateAndFillIndexBuffer(const unsigned int numIndice
 
 IndexBuffer* RenderDevice::CreateAndFillIndexBuffer(const std::vector<WORD>& indices)
 {
-   return CreateAndFillIndexBuffer(indices.size(), &indices[0]);
+   return CreateAndFillIndexBuffer((unsigned int)indices.size(), &indices[0]);
 }
 
 IndexBuffer* RenderDevice::CreateAndFillIndexBuffer(const std::vector<unsigned int>& indices)
 {
-   return CreateAndFillIndexBuffer(indices.size(), &indices[0]);
+   return CreateAndFillIndexBuffer((unsigned int)indices.size(), &indices[0]);
 }
