+++ propbrowser.cpp
@@ -341,7 +341,7 @@ void SmartBrowser::CreateFromDispatch(HWND hwndParent, Vector<ISelect> *pvsel)
    //for (int i=0;i<m_vhwndExpand.Size();i++) SendMessage(m_vhwndExpand.ElementAt(i), EXPANDO_EXPAND, 1, 0); 
 
    //expand top last
-   for (int i = m_vhwndExpand.size() - 1; i >= 0; --i)
+   for (int i = (int)m_vhwndExpand.size() - 1; i >= 0; --i)
       SendMessage(m_vhwndExpand[i], EXPANDO_EXPAND, 1, 0);
 
    for (int i = 0; i < m_pvsel->Size(); i++)
